Move OpentracingSpan to tracing.go
