+++ conf/defaults.ini
@@ -912,7 +912,6 @@ zipkin_propagation = false
 disable_shared_zipkin_spans = false
 
 [tracing.opentelemetry.jaeger]
-enabled =
 # jaeger destination (ex http://localhost:14268/api/traces)
 address =
